feat: rename pythin file, cchange setup function to use one connection, and remove document_matadata table check
